+++ NEWS.md
@@ -1,6 +1,8 @@
-# TreeTools 2.0.0.9003 (development) #
+# TreeTools 2.0.0.9004 (development) #
+- Support larger trees in `ClusterTable` objects.
+  * Retires `CT_PUSH` and `CT_POP` macros.
 - Support larger trees in `Consensus()`.
-  Uses 32-bit integers, necessitating downstream changes to TreeDist.
+  * Uses 32-bit integers, necessitating downstream changes to TreeDist.
 
 # TreeTools 2.0.0.9001 (development) #
 - Remove hard limit on tree size in `SplitList`.